SSH driver bug fix on buffer size.
To instantiate multiple variables, the pmacCommandStore::buildCommandString
function generates an addition of 40 variables in a 1024 char. However, the
SSHDriver::write function needs to perform an ECHO and the static buffer size
is 512 char. If the buffer size passed to SSHDriver::write is greater than
512 char, an error is generated in the function (no match), generating a false
error of connection loss and reconnection with the deltatau controller.
To solve the problem, the buffer size was changed to 2048 char.